compiler: Add local variables in the context of events
This patch adds a variable to track the current context (std::wstring,
"" means global, "event_xxx" for events, "sub_xxx" for subroutines).

"var xyz" statments are parsed at the begining of events, and
corresponding variables are allocated at the top of the memory
and added to a context-aware variables map.